+++ tests/chessboard_4px_raw.rs
@@ -1,4 +1,4 @@
-use tinytga::{Bpp, ImageOrigin, ImageType, RawTga, TgaHeader};
+use tinytga::{Bpp, ImageOrigin, DataType, Compression, RawTga, TgaHeader};
 
 #[test]
 fn chessboard_4px_raw() {
@@ -15,7 +15,8 @@ fn chessboard_4px_raw() {
         TgaHeader {
             id_len: 0,
             has_color_map: false,
-            image_type: ImageType::Truecolor,
+            data_type: DataType::TrueColor,
+            compression: Compression::Uncompressed,
             color_map_start: 0,
             color_map_len: 0,
             color_map_depth: None,